我的 团队管理会员返点显示小数点
(cherry picked from commit 196fb3964ac160a8922809422a1ca9cd4da6cd36)

+++ module-mine/src/main/java/com/xtree/mine/ui/fragment/MemberManagerAdapter.java
@@ -71,7 +71,7 @@ public class MemberManagerAdapter extends CachedAutoRefreshAdapter<MemberUserInf
 
         binding.tvwUserName.setText(vo.username);
         if (vo.userpoint != null) {
-            binding.tvwReturnPoint.setText((int) (Double.parseDouble(vo.userpoint) * 100) + "%");
+            binding.tvwReturnPoint.setText((Double.parseDouble(vo.userpoint) * 100) + "%");
             binding.tvwUserName.setEnabled(true); //名字不可点击
         } else {
             //binding.tvwReturnPoint.setText("0%");
